fix(mdm): preserve lease signing key continuity
